fix: rename auth reducer folder
